cmd/syncthing: Move HideConsole option stuff to cmdutil package.
