+++ cmd/syncthing/cmdutil/options_common.go
@@ -8,6 +8,7 @@ package cmdutil
 
 // CommonOptions are reused among several subcommands
 type CommonOptions struct {
+	buildCommonOptions
 	ConfDir         string `name:"config" placeholder:"PATH" help:"Set configuration directory (config and keys)"`
 	HomeDir         string `name:"home" placeholder:"PATH" help:"Set configuration and data directory"`
 	NoDefaultFolder bool   `env:"STNODEFAULTFOLDER" help:"Don't create the \"default\" folder on first startup"`
